+++ use-shopping-cart/.eslintrc
@@ -5,7 +5,8 @@
     "standard-react"
   ],
   "env": {
-    "es6": true
+    "es6": true,
+    "jest": true
   },
   "plugins": [
     "react"
