+++ use-shopping-cart/core/Entry.js
@@ -1,4 +1,4 @@
-import { formatCurrencyString } from '../src/util'
+import { formatCurrencyString } from '../utilities/old-utils'
 import { v4 as uuidv4 } from 'uuid'
 
 function Entry({
