+++ use-shopping-cart/core/store.js
@@ -1,8 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit'
-import { reducer } from '../core/slices/cartSlice'
+import { reducer, actions, initialState } from './slice'
 
-export default configureStore({
-  reducer: {
-    cart: reducer
-  }
-})
+export { reducer, actions }
+export function createShoppingCartStore(options) {
+  configureStore({
+    reducer,
+    preloadedState: { ...initialState, ...options }
+  })
+}
