+++ use-shopping-cart/react/index.d.ts
