+++ use-shopping-cart/utilities/SSR.js
@@ -1,0 +1,1 @@
+export const isClient = typeof window === 'object'
